refactor: make lib.rs clean and move all tests to integration tests
