+++ tests/api/main.rs
@@ -1,0 +1,2 @@
+mod bifurcate_coord;
+pub mod shared;
